Made some architectural changes to the system and documented some of the stub code so it's easier to go through.
The network code is divided by two concepts of:

 * A Network class
 * A NetworkCommunicator

The Network contains some of the main stuff. The NetworkCommunicator does most
of the actual work. The system is designed to start one or two communicators
(bad work.. have nothing to do with the MPI communicators.. will rethink the
name) depending on the user wanting one thread to do both incomming and outgoing
or one to handle them both. Now, to enable network type chainging in the future
the Network and NetworkCommunicator classes are implemented as abstract. Each
network should inherit these and build their own. Like we do in the TCP layer
with:

 * TCPNetwork
 * TCPNetworkCommunicator

The TCPNetwork is responsible for starting a listing socket and doing initial
structure communication with the mpirun script. After this the socket is passed
on to the tread handling the incomming requests. The networks job after this
startup is mainly to facilite jobs finished / started by the threads. That is,
communicate the changes to the MPI thread, either by receiving jobs or sending.

NOTE: This is not fully implemented yet. There are still some startup issues and
the main select part is not ready yet.