Implemented the Communicator.allreduce API call and wrote a small test for it calculating the fact function.
The generic broadcast tree was changed as it didn't know when the local data
should be used. The current solution is not enough I think, as the Collective
request will not be generic across ranks in the same operation. Example, in
broadcast only the root will use the local data. Everybody else will just
forward it.