Tried to make the BroadcastTree more printerfriendly by implementering a __repr__ method. Whenever a new tree is generated the system will print the tree so it's possible to debug how it's structured.
There are something of a minor bug with the all-to-all startup socket thing I
implemented. It's not blocking which we should find a way to make it. Right
now, a process can actually send and receive all it need and then start to
communicate with the other processes as it normally would. I'm unsure if this
could overlap with the startup communication. We need to draw something on
paper so make sure it's not the case.