MAKING USER ARGUMENTS WORK, by
Fixing the error, so the contents of sys.argv isn't looking like:

 [ /path/to/user/script.py [arg0, arg1, arg2] ]

byt the correct version:

 [ /path/to/user/script.py, arg0, arg1, arg2 ]

It's was simply a matter of calling extend() on a list instead of
append.

Created a test for testing user level argument. For this I added
some functionality in runtest.py so it's possible to enter user
level arguments by adding the #meta-userargs <your args here>.

This commit fixes #10