Reordered the tags in constants.py so they are not 2^n anymore. This will make it a bit easier to add tags in the future.
Implemented allgather and gather, which reduces the collective operations quite
a lot. The rest of the API will be completed in this weekend. Allgather and
gather have test files and end-user documentation.

NOTE: A this point the test will not always complete. I'm blaiming the shutdown
errors for this, but if not we need to make sure it's not the collective
operations failing.