Encountered a problem with the speedup calculation. We compared lists in the format [(datasize),(value)]. If there was not an equal amount of elements to compare the calculations would be invalid. For example, if we had 3 elements per data size for tag A and 4 for tag B we have the following situation:
 data[A][x1].val1 compared to data[B][x1].val1       correct
 data[A][x1].val2 compared to data[B][x1].val2       correct
 data[A][x1].val3 compared to data[B][x1].val3       correct
 data[A][x2].val1 compared to data[B][x1].val4       INCORRECT (and the rest)

The solution is to normalize the data so each list have the same elements for each x value. This means
we through away the slowest elements for the tag that has the most values.

Changed the option parser, so we can compare a single tag. This is just a quick way for us to draw performance plots for a single tag.

Due to our major speedup the axis formatting was wrong. We just used a simple python range(0, max),
but as we have speedup in the range of 300. Now, we make room for 20 labels, so in this example
we get:

   skip: 300 / 20 = 15
   labels: range(0, 300, 15) --> [0, 15, 30, 45, 60, .. 300]