This implements unix sockets as the main way to communicate between ranks on the same host. It is a draft so it is not tested in depths or performance benchmarked yet.
Each rank now creates two listining sockes: 1) A TCP socket listning on a random
port, 2) A unix socket listning on a tempfile. Access information for both (ie.
host, port and filename) is sent back to mpirun which distributes them. When the
collected information is sent back to each rank the list is traversed and only
the needed information is kept. That is, if the rank is on the same host as the
received we only save the filepath for the unix socket. Otherwise we juse the
TCP conncetion information.